- Add sample data - Minor style changes: renaming, whitespace, etc. - Add analyze sample data button - Add requirements.txt - Add some instructions in README - Improved portability to linux/mac - Refactor CSV loading to enable easy sample data analysis - add some TODO comments
