+++ requirements.txt
@@ -1,0 +1,8 @@
+# List third-party dependencies, pin exact versions for reproducibility
+# Python 3.9.*
+
+PyQt5==5.15.6
+numpy==1.17.0
+pandas==1.3.5
+pyqtgraph==0.12.3
+scikit-learn==1.0.2
